requirements.txt
pymongo[srv]
flask
requests
python-dotenv
bs4
tqdm
-e .
